Renames OC vaccine extract.
